Fix #142
